Seperates myItem and item for code cleanliness
