+++ frontEnd/src/components/item.jsx
@@ -1,7 +1,5 @@
 //Item framework
 import { useContext, useEffect, useRef } from "react";
-import useDeleteItem from "../utils/useDeleteItem";
-import useEditItem from "../utils/useEditItem";
 import useDistance from "../utils/useDistance";
 import { Context } from "../App";
 import usePreferredMeetLocation from "../utils/usePreferredMeetLocation";
@@ -9,19 +7,10 @@ import usePreferredMeetLocation from "../utils/usePreferredMeetLocation";
 function Item({ postType, userId, isMyPost, item, onPostChange }) {
   const itemRef = useRef(null);
   const signedInUserAddress = useContext(Context).user.address;
-  const { fetchDelete } = useDeleteItem(userId, postType, item.id);
-  const { fetchEdit } = useEditItem(userId, postType, item.id);
   const { fetchDistance, distance, errorMsg } = useDistance();
 
   const { fetchPreferredMeetLocation, meetLocation } =
     usePreferredMeetLocation(userId);
-
-  /*This function is called when we want to delete an item from the list*/
-  const deleteItem = async (event) => {
-    event.preventDefault();
-    await fetchDelete();
-    onPostChange();
-  };
 
   useEffect(() => {
     if (signedInUserAddress && meetLocation) {
@@ -37,35 +26,6 @@ function Item({ postType, userId, isMyPost, item, onPostChange }) {
   useEffect(() => {
     fetchPreferredMeetLocation();
   }, []);
-  /*function to edit an item, will grab certain values and send a put to the server */
-  const postItemEdits = async (event) => {
-    event.preventDefault();
-    const parentItem = itemRef.current;
-    const description = parentItem.querySelector("#description").value;
-    const title = parentItem.querySelector("#title").value;
-    const itemState = parentItem.querySelector("#useStates").value;
-
-    if (title !== "" && description !== "") {
-      fetchEdit(
-        JSON.stringify({
-          id: item.id,
-          title: title,
-          description: description,
-          photo: "",
-          itemState: itemState,
-          userId: userId,
-        })
-      );
-      parentItem.querySelector("#title").placeholder =
-        parentItem.querySelector("#title").value;
-      parentItem.querySelector("#title").value = "";
-      parentItem.querySelector("#description").placeholder =
-        parentItem.querySelector("#description").value;
-      parentItem.querySelector("#description").value = "";
-    } else {
-      alert("Missing Title or Desciption");
-    }
-  };
 
   return (
     <div
@@ -73,65 +33,34 @@ function Item({ postType, userId, isMyPost, item, onPostChange }) {
       ref={itemRef}
     >
       <div style={{ width: "44%", marginLeft: "6%" }}>
-        {isMyPost ? (
-          <div>
-            <div>
-              <label htmlFor="title">Title: </label>
-              <input
-                id="title"
-                name="title"
-                type="text"
-                placeholder={item.title}
-              ></input>
-            </div>
-            <label htmlFor="description">Description: </label>
-            <input
-              name="description"
-              id="description"
-              type="text"
-              placeholder={item.description}
-            ></input>
-            <br></br>
-            <select name="useState" id="useStates" defaultValue={item.useState}>
-              <option value="Used Like New">Used Like New</option>
-              <option value="Used">Used</option>
-              <option value="New">New</option>
-            </select>
-            <div>
-              <button onClick={postItemEdits}>Submit Edits</button>
-              <button onClick={deleteItem}>Delete</button>
-            </div>
-          </div>
-        ) : (
-          <div>
-            <h2>{item.title}</h2>
-            <p>{item.description}</p>
-            <p>Use State: {item.itemState}</p>
-            <p>Distance: {distance ? distance : errorMsg}</p>
-            {postType === "donations" && (
-              <form>
-                <label style={{ marginLeft: "10px" }} htmlFor="wantScore">
-                  Want Score:{" "}
-                </label>
-                <select name="wantScore">
-                  <option value="1">1</option>
-                  <option value="2">2</option>
-                  <option value="3">3</option>
-                  <option value="4">4</option>
-                  <option value="5">5</option>
-                  <option value="6">6</option>
-                  <option value="7">7</option>
-                  <option value="8">8</option>
-                  <option value="9">9</option>
-                  <option value="10">10</option>
-                </select>
-                <button style={{ marginTop: "10px", marginBottom: "10px" }}>
-                  Request Item
-                </button>
-              </form>
-            )}
-          </div>
-        )}
+        <div>
+          <h2>{item.title}</h2>
+          <p>{item.description}</p>
+          <p>Use State: {item.itemState}</p>
+          <p>Distance: {distance ? distance : errorMsg}</p>
+          {postType === "donations" && (
+            <form>
+              <label style={{ marginLeft: "10px" }} htmlFor="wantScore">
+                Want Score:{" "}
+              </label>
+              <select name="wantScore">
+                <option value="1">1</option>
+                <option value="2">2</option>
+                <option value="3">3</option>
+                <option value="4">4</option>
+                <option value="5">5</option>
+                <option value="6">6</option>
+                <option value="7">7</option>
+                <option value="8">8</option>
+                <option value="9">9</option>
+                <option value="10">10</option>
+              </select>
+              <button style={{ marginTop: "10px", marginBottom: "10px" }}>
+                Request Item
+              </button>
+            </form>
+          )}
+        </div>
       </div>
       <div style={{ width: "44%" }}>
         <img
